app/rag/requirements.txt
azure-identity==1.19.0
azure-storage-blob==12.23.1
langchain-text-splitters==0.3.2
langfuse==2.53.0
openai==1.55.0
pgvector==0.3.6
psycopg[binary]==3.2.3
pypdf==5.1.0
python-docx==1.1.2
structlog==24.4.0
